Auto-onboard OIDC users API calls
When an API call is made with ID token as Bearer token, but
no according user exists yet, the user is onboarded automatically
based on the ID token claims.

Prerequisite is, that auto-onboarding is enabled in OIDC settings.
Also, a valid username must be provided in configured username
claim or default "name" claim.

Signed-off-by: Steffen Templin <steffen.templin@advantest.com>